Lower string-typed + to the target concat operator
JS overloads + on operand type; Perl EP, Kolon, Twig, and Blade all
have numeric-only + — 'Hello, ' + name rendered 0 on the Perl family
and fataled on PHP ("Unsupported operand types"). The string-typed
classification is shared-layer semantics: isStringTypedOperand
(promoted from the Mojo/Xslate adapters' local copies, extended with
template-literal and nested-+ arms) and isStringConcatBinary are now
exported from @barefootjs/jsx, and each emitter only maps the shared
decision to its own operator (EP/Blade `.`, Kolon/Twig `~`).

The string-concat-plus fixture graduates from those four adapters'
renderDivergences (Jinja/minijinja/ERB concatenate natively; the Go
adapter lowers through its own pipeline and keeps its entry).
