+++ packages/adapter-mojolicious/src/adapter/expr/operand.ts
@@ -1,39 +1,18 @@
 /**
- * Operand-type classification + index-access lowering for the Mojolicious
- * EP template adapter.
+ * Index-access lowering for the Mojolicious EP template adapter.
  *
  * Extracted from `mojo-adapter.ts` (domain-module refactor, issue #2018
- * track D). Pure functions over `ParsedExpr` — they take an `isStringName`
+ * track D). Pure function over `ParsedExpr` — it takes an `isStringName`
  * predicate (supplied by the emitter from adapter state) rather than reading
  * adapter instance state directly.
  *
- * SHARED CANDIDATE: `isStringTypedOperand` is byte-identical to the Xslate
- * adapter's copy and is adapter-agnostic — an extraction candidate for a
- * shared Perl-family codegen module (groundwork for the future Perl evaluator
- * integration, issue #2018 track D). `emitIndexAccessPerl` stays Mojo-specific
+ * The string-typed-operand classifier that used to live here (marked
+ * SHARED CANDIDATE) was promoted to `@barefootjs/jsx` as
+ * `isStringTypedOperand` (#2176); `emitIndexAccessPerl` stays Mojo-specific
  * (Perl's `->[]` vs `->{}` split has no Kolon equivalent).
  */
 
-import type { ParsedExpr } from '@barefootjs/jsx'
-
-/**
- * Whether a comparison operand is string-typed, so JS `===`/`!==` against it
- * must lower to Perl `eq`/`ne` instead of numeric `==`/`!=` (#1672). Covers a
- * string literal, a string-signal getter call (`sel()`), and a string prop
- * access (`props.x`). `isStringName` reports whether a getter/prop name is
- * known-string. Loop-element fields (`t.id`) on untyped arrays have no known
- * type and stay undetected — a separate, narrower gap.
- */
-export function isStringTypedOperand(expr: ParsedExpr, isStringName: (n: string) => boolean): boolean {
-  if (expr.kind === 'literal' && expr.literalType === 'string') return true
-  if (expr.kind === 'call' && expr.callee.kind === 'identifier' && expr.args.length === 0) {
-    return isStringName(expr.callee.name)
-  }
-  if (expr.kind === 'member' && expr.object.kind === 'identifier' && expr.object.name === 'props') {
-    return isStringName(expr.property)
-  }
-  return false
-}
+import { isStringTypedOperand, type ParsedExpr } from '@barefootjs/jsx'
 
 /**
  * Lower `arr[index]` to a Perl deref. Perl distinguishes array
